requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
PyYAML
pytest
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
